+++ ExplorePackages/Entities/PackageArchiveEntity.cs
@@ -4,9 +4,9 @@
     {
         public long PackageKey { get; set; }
         public long Size { get; set; }
-        public long EntryCount { get; set; }
-        public long OffsetOfCentralDirectory { get; set; }
-        public long? Zip64OffsetOfCentralDirectory { get; set; }
+        public int EntryCount { get; set; }
+        public uint OffsetOfCentralDirectory { get; set; }
+        public ulong? Zip64OffsetOfCentralDirectory { get; set; }
 
         public PackageEntity Package { get; set; }
     }
